+++ R/margot_batch_policy.R
@@ -8,7 +8,7 @@
 #' @param height The height of saved plots in inches. Default is 12.
 #' @param save_plots Logical indicating whether to save plots to disk. Default is TRUE.
 #' @param output_dir The directory to save plots in. Default is here::here(push_mods).
-#' @param spend_levels A vector of spend levels to use for difference gain summaries. Default is c(0.2, 0.5).
+#' @param spend_levels A vector of spend levels to use for difference gain summaries. Default is c(0.1, 0.4).
 #' @param label_mapping Optional named list for custom label mappings. Keys should be original variable names
 #'        (with or without "model_" prefix), and values should be the desired display labels. Default is NULL.
 #' @return A list where each element corresponds to a model in the input
@@ -33,7 +33,7 @@ margot_batch_policy <- function(result_outcomes,
                                 height = 12,
                                 save_plots = TRUE,
                                 output_dir = here::here(push_mods),
-                                spend_levels = c(0.2, 0.5),
+                                spend_levels = c(0.1, 0.4),
                                 label_mapping = NULL) {
   # Deprecation warning
   warning("The margot_batch_policy() function is deprecated as of margot 0.2.1.65. Please use margot_policy() instead.",